analysis-worker/requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
google-cloud-storage
earthengine-api
numpy
Pillow
# opencv-python-headless
# torch
# torchvision
# segment-anything
google-auth